+++ schemasync-backend/src/main/java/com/schemasync/adapter/TDSQLAdapter.java
@@ -176,17 +176,18 @@ public class TDSQLAdapter implements DatabaseAdapter {
                     column.setColumnName(rs.getString("COLUMN_NAME"));
                     column.setDataType(rs.getString("DATA_TYPE"));
                     
-                    int length = rs.getInt("CHARACTER_MAXIMUM_LENGTH");
+                    // 使用Long支持超大值
+                    long length = rs.getLong("CHARACTER_MAXIMUM_LENGTH");
                     if (!rs.wasNull()) {
                         column.setLength(length);
                     }
                     
-                    int precision = rs.getInt("NUMERIC_PRECISION");
+                    long precision = rs.getLong("NUMERIC_PRECISION");
                     if (!rs.wasNull()) {
                         column.setPrecision(precision);
                     }
                     
-                    int scale = rs.getInt("NUMERIC_SCALE");
+                    long scale = rs.getLong("NUMERIC_SCALE");
                     if (!rs.wasNull()) {
                         column.setScale(scale);
                     }
